Created functions for assignment 2
I created the makeCacheMatrix and cacheSolve functions.  makeCacheMatrix
creates and returns an R object that stores a matrix and caches the
inverse.  The cacheSolve function takes a cacheMatrix object and returns
the inverse of the matrix stored in the object, either by retrieving the
cached inverse matrix or calculating it.  If the inverse matrix is
calculated, it is also cached in the cacheMatrix object.